+++ NEWS.md
@@ -6,9 +6,14 @@
 
 * Update `drug_regimen_list` lookup table to include drug names by data release, as opposed to by cohort. It is possible that drug names were modified slightly across data releases (#132)
 
+* NEW FEATURE: Users can now pass their Personal Access Token (as well as username/ password) to set up authentication for accessing all data. (#119)
+
+* BREAKING CHANGE: Users must use `set_synapse_credentials()` at the top of all scripts and explicitly set username/password or PAT for each session.
+
 * Added **NSCLC v3.1-consortium** release (#137)
 
 * Read in structural variant files for NSCLC v2.2-consortium, CRC v1.3-consortium, and Bladder v1.2-consortium. For more information on structural variant data, see: [https://docs.cbioportal.org/file-formats/#structural-variant-data](https://docs.cbioportal.org/file-formats/#structural-variant-data). 
+
 
 # genieBPC 1.1.1